pyproject.toml
[project]
name = "resumaster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.45.2",
    "fastapi>=0.115.8",
    "html2text>=2024.2.26",
    "jinja2>=3.1.5",
    "logfire[fastapi,system-metrics]>=3.4.0",
    "markdown>=3.7",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.0",
]
